rf-15: validateRoutineCreation middleware refactor
