+++ src/routes/routine/index.ts
@@ -1,11 +1,13 @@
 import express from 'express';
 
+import { validateRoutineCreation } from 'src/validations/routine';
+
 import controllers from './controllers';
 
 const router = express.Router();
 
 router.get('/', controllers.getAllRoutines);
-router.post('/', controllers.createRoutine);
+router.post('/', validateRoutineCreation, controllers.createRoutine);
 router.put('/:id', controllers.editRoutine);
 router.delete('/:id', controllers.deleteRoutine);
